move chromatogram file writing to module
